requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-htmx
crispy-bootstrap5==2024.10
    # via -r requirements.in
django==4.2.17
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-ckeditor
    #   django-crispy-forms
    #   django-dynamic-breadcrumbs
    #   django-htmx
    #   django-js-asset
django-ckeditor==6.7.2
    # via -r requirements.in
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-dynamic-breadcrumbs==0.4.3
    # via -r requirements.in
django-htmx==1.21.0
    # via -r requirements.in
django-js-asset==2.2.0
    # via django-ckeditor
gunicorn==23.0.0
    # via -r requirements.in
packaging==24.2
    # via gunicorn
pillow==11.0.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via asgiref
tzdata==2024.2
    # via django
